+++ addons/l10n_lu/demo/demo_company.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
     <record id="partner_demo_company_lu" model="res.partner">
-        <field name="name">LU Company</field>
+        <field name="name">My Luxembourg Company</field>
         <field name="vat">LU75425064</field>
         <field name="street">A</field>
         <field name="city">Clervaux</field>
@@ -14,7 +14,7 @@
     </record>
 
     <record id="demo_company_lu" model="res.company">
-        <field name="name">LU Company</field>
+        <field name="name">My Luxembourg Company</field>
         <field name="partner_id" ref="partner_demo_company_lu"/>
     </record>
 
@@ -24,12 +24,6 @@
 
     <function model="res.users" name="write">
         <value eval="[ref('base.user_root'), ref('base.user_admin'), ref('base.user_demo')]"/>
-        <value eval="{'company_ids': [(4, ref('l10n_lu.demo_company_lu'))]}"/>
-    </function>
-
-    <function model="account.chart.template" name="try_loading">
-        <value eval="[]"/>
-        <value>lu</value>
-        <value model="res.company" eval="obj().env.ref('l10n_lu.demo_company_lu')"/>
+        <value eval="{'company_ids': [(4, ref('l10n_lu.demo_company_lu'))], 'tz': 'Europe/Luxembourg'}"/>
     </function>
 </odoo>
