+++ addons/l10n_lu_account/i18n_extra/l10n_lu.pot
